pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-architecture-patterns"
version = "0.1.0"
description = "Companion repo for the RAG Architecture Patterns blog post"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
dependencies = []

[tool.setuptools]
# Explicit package list — flat-layout repo with multiple top-level dirs (shared, scripts, dataset, evaluation, tier-*).
# Only the Python source packages are installed; dataset/evaluation/tier-* are not Python packages.
# ``tier_1_naive`` and ``tier_3_graph`` are importable shim packages that re-export
# modules from the hyphenated ``tier-1-naive/`` / ``tier-3-graph/`` directories
# (hyphens are not valid Python identifiers).
packages = ["shared", "scripts", "tier_1_naive", "tier_3_graph"]

[project.optional-dependencies]
shared = [
  "google-genai>=1.73,<2",
  "pydantic>=2.10,<3",
  "pydantic-settings>=2.10,<3",
  "rich>=14,<16",
  "tiktoken>=0.10,<1",
  "python-dotenv>=1.0,<2",
]
curation = [
  "arxiv>=3.0,<4",
  "semanticscholar>=0.12,<1",
  "pymupdf>=1.27,<2",
  "Pillow>=10,<12",
]
tier-1 = [
  "rag-architecture-patterns[shared]",
  "chromadb>=1.5.8,<2",
  "openai>=1.50,<3",
  "pymupdf>=1.27,<2",
]
tier-2 = ["rag-architecture-patterns[shared]"]
tier-3 = [
  "rag-architecture-patterns[shared]",
  "lightrag-hku==1.4.15",
  "openai>=1.50,<3",
  "pymupdf>=1.27,<2",
]
tier-4 = ["rag-architecture-patterns[shared]"]
tier-5 = ["rag-architecture-patterns[shared]"]

[dependency-groups]
test = ["pytest>=8.4,<9"]
dev = [{include-group = "test"}, "ruff>=0.6"]

[tool.uv]
default-groups = ["dev"]

[tool.pytest.ini_options]
markers = ["live: tests that hit real APIs and incur cost"]
